// types.ts
export type PrismicRichText = any;

export type Homepage = {
  title?: PrismicRichText;
  intro?: PrismicRichText;
  link?: PrismicRichText;
  link2?: PrismicRichText;
};

export type Slice = {
  context?: PrismicRichText;
  img?: PrismicRichText;
  date?: PrismicRichText;
  color?: PrismicRichText;
};